Feature/request mapping

+++ src/main/java/edu/makarov/customer/service/impl/AccountServiceImpl.java
@@ -1,6 +1,7 @@
 package edu.makarov.customer.service.impl;
 
 import edu.makarov.customer.models.Account;
+import edu.makarov.customer.models.Customer;
 import edu.makarov.customer.repository.AccountRepository;
 import edu.makarov.customer.service.AccountService;
 import edu.makarov.customer.service.CustomerService;
@@ -13,11 +14,14 @@ import java.util.List;
 @Service
 public class AccountServiceImpl implements AccountService {
 
-    @Autowired
-    private AccountRepository accountRepository;
+    private final AccountRepository accountRepository;
+    private final CustomerService customerService;
 
     @Autowired
-    private CustomerService customerService;
+    public AccountServiceImpl(AccountRepository accountRepository, CustomerService customerService) {
+        this.accountRepository = accountRepository;
+        this.customerService = customerService;
+    }
 
     @Override
     public List<Account> findAll() {
@@ -26,12 +30,15 @@ public class AccountServiceImpl implements AccountService {
 
     @Override
     public List<Account> findAllByCustomerId(long id) {
+        if (customerService.findById(id) == null) {
+            return null;
+        }
         return accountRepository.findAccountByCustomerId(id);
     }
 
     @Override
     public Account findById(long id) {
-        return accountRepository.findById(id).get();
+        return accountRepository.findById(id).orElse(null);
     }
 
     @Override
@@ -41,19 +48,30 @@ public class AccountServiceImpl implements AccountService {
 
     @Override
     public Account create(Account account, long customerId) {
-        account.setCustomer(customerService.findById(customerId));
+        Customer customer = customerService.findById(customerId);
+        if (customer == null) {
+            return null;
+        }
+        account.setCustomer(customer);
         return create(account);
     }
 
     @Override
     public Account update(long id, Account account) {
         Account accountFromDb = findById(id);
+        if (accountFromDb == null) {
+            return null;
+        }
         BeanUtils.copyProperties(account, accountFromDb, "id");
         return create(accountFromDb);
     }
 
     @Override
-    public void delete(long id) {
+    public boolean delete(long id) {
+        if (findById(id) == null) {
+            return false;
+        }
         accountRepository.deleteById(id);
+        return true;
     }
 }
